# openspec/steel-marketplace-docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Steel Marketplace API
  version: "1.0.0"
  description: |
    REST API for a B2B marketplace platform that supports:
      - Goods procurement (RFQ/offer/selection) for steel products
      - Industrial services (cutting, shipping, quality inspection)
      - Composite orders (parent/child)
      - Milestone payments with escrow-like holding and final weight adjustment

    Backend: Django + Django REST Framework.
servers:
  - url: https://api.example.com/api/v1
    description: Production (example)
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders visible to the current user
      description: |
        - Buyers: their own orders.
        - Providers: orders assigned to them + "available marketplace feed" orders
          (depending on endpoint policy; many teams implement a dedicated /marketplace/orders).
        - Admin: all orders.

        Use query params to filter by type/status/parent_order_id.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: order_type
          in: query
          schema:
            $ref: "#/components/schemas/OrderType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: parent_order_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
    post:
      tags: [Orders]
      summary: Create an order (BUY/CUT/SHIP/QC/COMBO)
      description: |
        Creates a new order. For COMBO orders, the server may create child orders in one call
        if `children` are provided.

        Note: Providers typically cannot create BUY orders as buyers; role checks apply.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
            examples:
              buy_order:
                value:
                  order_type: BUY
                  title: "Buy ~10 tons steel sheet"
                  product_type_id: "9b84a7c2-9c6a-4b22-8a1a-9f4d2f6e0d6e"
                  material_spec:
                    grade: "ST37"
                    thickness_mm: 10
                    width_mm: 1500
                    length_mm: 6000
                  quantity:
                    unit: TON
                    estimated_value: 10.0
                    tolerance_percent: 2.0
                  delivery:
                    city: "Tehran"
                    address: "Project site address..."
                    window_start: "2026-01-10T08:00:00Z"
                    window_end: "2026-01-15T18:00:00Z"
                  requested_services: ["SHIP", "QC"]
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /marketplace/feed/:
    get:
      tags: [Order Requests]
      summary: List marketplace request feed
      description: |
        Returns only active requests:
        - BUY requests (active)
        - SELL requests with warehouse status APPROVED
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/OrderRequestType"
        - name: category
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderRequestStatus"
      responses:
        "200":
          description: Marketplace feed list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrderRequestFeedItems"
        "403":
          $ref: "#/components/responses/Forbidden"
  /my/requests/:
    get:
      tags: [Order Requests]
      summary: List my request history
      description: Includes active and deactivated requests for the current user.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: My requests list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrderRequests"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Order Requests]
      summary: Create BUY or SELL request
      description: |
        - BUY requires active BUYER role and starts as APPROVED.
        - SELL requires active SELLER role and starts as PENDING_WAREHOUSE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestCreateRequest"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /my/requests/{request_id}/:
    get:
      tags: [Order Requests]
      summary: Get my request details
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Order Requests]
      summary: Update or deactivate my request
      description: |
        Use `is_active=false` to soft-delete (deactivate) request.
        Material edits on an approved SELL request reset status to PENDING_WAREHOUSE.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestUpdateRequest"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Order Requests]
      summary: Deactivate my request (soft-delete)
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "204":
          description: Request deactivated
        "404":
          $ref: "#/components/responses/NotFound"
  /my/requests/{request_id}/documents/:
    post:
      tags: [Order Requests]
      summary: Upload request documents
      description: "Accepts one or multiple files. Allowed file types: PDF, ZIP."
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Documents uploaded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderRequestDocument"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /warehouse/verify/{request_id}/:
    post:
      tags: [Order Requests]
      summary: Approve or reject SELL request by warehouse manager
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseVerificationRequest"
      responses:
        "200":
          description: Request verification status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /orders/{order_id}/offers:
    get:
      tags: [Offers]
      summary: List offers for an order
      description: |
        - Buyer: sees all offers for their order.
        - Providers: see only their own offers for that order.
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Offers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOffers"
    post:
      tags: [Offers]
      summary: Submit an offer for an order (provider)
      description: |
        Provider submits an offer. Role must match the order_type:
          - BUY -> SELLER
          - CUT -> CUTTER
          - SHIP -> DRIVER
          - QC  -> QC_EXPERT
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferCreateRequest"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    OrderRequestType:
      type: string
      enum: [BUY, SELL]
...
    OrderType:
      type: string
      enum: [BUY, CUT, SHIP, QC, COMBO]
    OrderStatus:
      type: string
      description: |
        High-level statuses. Some order types may not use all statuses.
      enum:
        - POSTED
        - OFFER_SELECTED
        - DEPOSIT_PENDING
        - DEPOSIT_PAID
        - SELLER_CONFIRM_PENDING
        - CONFIRMED
        - PAYMENT_IN_PROGRESS
        - READY_FOR_LOADING
        - LOADING
        - WEIGHING_PENDING
        - FINAL_PAYMENT_PENDING
        - FINALIZED
        - DRIVER_ASSIGNED
        - PICKUP_PENDING
        - IN_TRANSIT
        - DELIVERED
        - REPORT_SUBMITTED
        - COMPLETED
        - CANCELLED
        - EXPIRED
        - DISPUTED
...
    Order:
      type: object
      required: [id, order_type, status, buyer_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        order_type:
          $ref: "#/components/schemas/OrderType"
        status:
          $ref: "#/components/schemas/OrderStatus"
        title:
          type: string
          nullable: true
        buyer_id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
          nullable: true
          description: Assigned provider (seller/cutter/driver/qc) after selection.
        parent_order_id:
          type: string
          format: uuid
          nullable: true
        product_type_id:
          type: string
          format: uuid
          nullable: true
        material_spec:
          $ref: "#/components/schemas/MaterialSpec"
        quantity:
          $ref: "#/components/schemas/Quantity"
        delivery:
          $ref: "#/components/schemas/DeliveryWindow"
        requested_services:
          type: array
          items:
            type: string
            enum: [CUT, SHIP, QC]
          description: For BUY or COMBO orders.
        chosen_offer_id:
          type: string
          format: uuid
          nullable: true
        pricing:
          type: object
          nullable: true
          properties:
            unit_price:
              $ref: "#/components/schemas/Money"
            total_estimated:
              $ref: "#/components/schemas/Money"
            total_final:
              $ref: "#/components/schemas/Money"
        weight:
          type: object
          nullable: true
          properties:
            estimated:
              type: number
              format: float
              nullable: true
            actual:
              type: number
              format: float
              nullable: true
            unit:
              type: string
              enum: [KG, TON]
              nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    OrderCreateRequest:
      type: object
      required: [order_type]
      properties:
        order_type:
          $ref: "#/components/schemas/OrderType"
        title:
          type: string
        product_type_id:
          type: string
          format: uuid
        material_spec:
          $ref: "#/components/schemas/MaterialSpec"
        quantity:
          $ref: "#/components/schemas/Quantity"
        delivery:
          $ref: "#/components/schemas/DeliveryWindow"
        requested_services:
          type: array
          items:
            type: string
            enum: [CUT, SHIP, QC]
        parent_order_id:
          type: string
          format: uuid
          nullable: true
        children:
          type: array
          description: Optional child orders for COMBO creation.
          items:
            type: object
            properties:
              order_type:
                $ref: "#/components/schemas/OrderType"
              title:
                type: string
              material_spec:
                $ref: "#/components/schemas/MaterialSpec"
              quantity:
                $ref: "#/components/schemas/Quantity"
              delivery:
                $ref: "#/components/schemas/DeliveryWindow"
...
    PaginatedOrders:
      type: object
      required: [count, results]
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/Order"
...